refactor: cleaning DevOpUtils/UninstallTemplate.Program.cs #85 #89
